Grid: Round offset data to handle high-dpi monitors
When labels are rendered in the grid overlay, they are positioned thanks to
offsets that the backend calculates. These offsets may be floats when the
browser displays content on a high-dpi monitor.
The code uses these offsets to make decisions like: "does this label correspond
to the last line in the grid". Doing this require comparing label offsets and
grid container sizes, which can only be done correctly if we round all the
involved values to the nearest integer

If we don't do this, things like shifting labels a few pixels when they're
displayed at an inner corner of the grid won't work.

Bug: 1098929
Change-Id: I9a210b462bed9ccd73eafd5f7ae991dffed27139