Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> I have honestly no idea if this is the right solution. Could you look into the two layout test failures? They seem related to the changes. Let's take it from there, after the tests are fixed, maybe you uncover another rabbit hole.

They are related and they will need to be disabled to check this in. One of them is simply due to expecting a snippet to be named with %escapes and the test will need to be fixed. The automapping test that Jan added I am less certain about. It appears to be testing against the local filesystem, but I am uncertain how to duplicate what it's doing either in manual testing or in an e2e test. Jan, can you shed any light on what that test is doing and how I might try to reproduce it manually?

Patch-set: 4
Label: Commit-Queue=0