Update patch set 1
Patch Set 1: Code-Review+1

> Patch Set 1: Auto-Submit+1
> 
> We ran into this with https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2268246
> 
> This being a default for stylelint seems really dumb to me, but... so it goes.

Agreed that it's weird for stylelint to hard-error on this by default instead of just logging a warning.

The only downside to this change is that at some point our globbing might fail and expand to 0 files (maybe only on certain platforms?) and we wouldn't know about it. But at this point we know it works properly on all bots, so it's probably fine...

Patch-set: 1
Reviewer: Gerrit User 1214214 <1214214@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+1