Grid: Show grid icon in overlay tooltip also without grid highlight info
In the previous implementation, the grid icon in the element tooltip of
the overlay would only be shown if the grid highlight config information
was being sent by the backend.
We don't want to have this restriction and instead need to be able to
show the grid icon at all times.

This change builds on a backend change whereby the type of layout object
currently being highlighted is passed as part of the elementInfo that
the backend already sends to the overlay frontend.
With this backend change, all highlighted elements come with a layout
type. We decided to only use this information for grids for now, but in
the future, it will be very easy to modify the code here to also show
icons for flex containers, table, inline, blocks, etc.

Corresponding backend CL: https://chromium-review.googlesource.com/c/chromium/src/+/2266385/

Bug: 1047356
Change-Id: Ia4521797c56a936f178fdc1bd9526015029a3764