Introduce the PageResourceLoader bottleneck for loading resources
This introduces a singleton PageResourceLoader and refactors loading of
additional developer resources to go through the loader. The loader
keeps track of all resources that get loaded, together with potential
load errors. If we ever decide to use a different method for loading
resources, we only need to change the PageResourceLoader.

Additionally, the loader limits the number of concurrent requests to
500, so as to not overload the backend. This is important, because
there is a limit of 2700 concurrent requests in the network stack -
if that limit is reached, additional requests will fail. Previously,
we introduced code in the back-end that retries requests, but this is
not a good approach, because only DevTools requests are retried.

Doing the rate limiting in the front-end has the downside that multiple
DevTools sessions don't have a shared limit, but simplifies the require-
ments for the back-end, so we think it is a good trade-off.

Bug: chromium:1069378
Change-Id: I05a9f3c79c685347c1e48b30c49f56801ae39fc3