Update patch set 1
Patch Set 1:

@Tim what's the difference / relationship between this CL and https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2256989?

Also, should we be worried about having some code that is minified by Terser, and a lot of other code that is minified by the legacy Python tool? I don't think it's an issue (we know our long term goal is to move to Terser 100%) but wanted to raise the question.

Patch-set: 1