Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> @Tim what's the difference / relationship between this CL and https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2256989?
> 
> Also, should we be worried about having some code that is minified by Terser, and a lot of other code that is minified by the legacy Python tool? I don't think it's an issue (we know our long term goal is to move to Terser 100%) but wanted to raise the question.

The difference is that the linked CL is a bit safer to perform, as well as it doesn't have the performance penalty for building in Chromium. Since `build_release_applications.py` is a monolithic task, including `terser` incurs a large build performance cost.

Yes, that does mean we use different minification compared to old code. Given that the impact of terser might be bigger than expected, I think that is a safer route for now.

Patch-set: 1