Update patch set 5
Patch Set 5: Code-Review-1

(1 comment)

(Putting a -1 here not because I will infinitely block this but to make sure we chat about this and figure out the best path forwards). 

I'm quite strongly opposed to the custom rollup plugin and overloading ESM imports - I think if we do want to go down this road we need to explore this thoroughly and probably share a design doc where we consider the impact of this and if it is the right path to go down.

Patch-set: 5
Reviewer: Gerrit User 1396442 <1396442@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=-1