Delay rendering of box model until style sections are initialized
Currently in the Styles pane, the box model (MetricsSidebarPane) renders
before the CSS properties sections, which creates a visual flash and
may confuse developers. This CL fixes this flash by delaying the
rendering of the box model until the style sections are initialized.

screencasts: https://imgur.com/a/2v6AyIw

Bug: chromium:1095547
Change-Id: Ib201e7c9ebf96f41ba96423ff830a71332c6fcc4