Update npm scripts to explicitly call python executable
This CL fixes an issue on Windows only where some npm scripts which
invoke the node.py script are not running, because cmd does not
understand the hashbang `#!/usr/bin/env python`.

This updates the calls to node.py such that they are all prefixed
with a call to python, following the existing pattern in the scripts,
which should work on all platforms.

This CL also fixes a small bug with the newly introduced "watch" script
where it was not producing output by moving the `--output` flag to the
first argument, a requirement of the node.py script.

Bug: 1100687
Change-Id: I003fb72b91f68febaf67b290e6836b8df33cdf41