[Loc] Fix <LOCKED> and <ex> get removed when messages move acrosss grdps
When the presubmit moves a message from one grdp file to another, the <ph name=LOCKED> and <ex> tags are lost because the messages are re-created from the frontend version of the string.
This change preserve the grd version of the strings that contains all the tags and use it when the message is moving across different grdp files.

Bug: 1071004
Change-Id: I5c724f0dc025dbfd06963373a8cb5fe7caf52803