Update patch set 6
Patch Set 6:

Jack and I discussed the constraints of the overlay in a VC and I documented them explicitly in the README now and made it clear that the custom rollup plugin should not be used by DevTools and it's specific to the overlay implementation. Also, I changed the CSS modules to return a constructed stylesheet that needs to be explicitly adopted in the code instead of relying on a side effect. PTAL!

Patch-set: 6