Add developer resources tab (behind a flag)
This ads a tab where all developer resources (that is source maps
and source code referenced from them) are listed once a load is
triggered. The tabs main function is to display the network load
status of the resources: It lists whether the load was successful
or not, and provides an error description. This is useful for
debugging when we enable a different source map loading mechanism.

The tab is currently gated behind a flag.

Bug: chromium:1069378
Change-Id: I84e880cc45c6520066d12310aff119fee5749fe1