Update patch set 8
Patch Set 8:

> Patch Set 6:
> 
> Jack and I discussed the constraints of the overlay in a VC and I documented them explicitly in the README now and made it clear that the custom rollup plugin should not be used by DevTools and it's specific to the overlay implementation. Also, I changed the CSS modules to return a constructed stylesheet that needs to be explicitly adopted in the code instead of relying on a side effect. PTAL!

If we are to introduce different infrastructure for the overlay than the rest of the `/front_end/` codebase, I think that warrants moving the code out of `/front_end/` into `/inspector_overlay/`. At that point, it is more reasonable to introduce custom stuff for overlays. Right now, we are mixing and matching responsibilities and it becomes harder to maintain the overall infrastructure of `/front_end/`.

Patch-set: 8