Update patch set 8
Patch Set 8:

> Patch Set 8:
> 
> > Patch Set 8:
> > 
> > > Patch Set 8:
> > > 
> > > > Patch Set 8:
> > > > 
> > > > > Patch Set 6:
> > > > > 
> > > > > Jack and I discussed the constraints of the overlay in a VC and I documented them explicitly in the README now and made it clear that the custom rollup plugin should not be used by DevTools and it's specific to the overlay implementation. Also, I changed the CSS modules to return a constructed stylesheet that needs to be explicitly adopted in the code instead of relying on a side effect. PTAL!
> > > > 
> > > > If we are to introduce different infrastructure for the overlay than the rest of the `/front_end/` codebase, I think that warrants moving the code out of `/front_end/` into `/inspector_overlay/`. At that point, it is more reasonable to introduce custom stuff for overlays. Right now, we are mixing and matching responsibilities and it becomes harder to maintain the overall infrastructure of `/front_end/`.
> > > 
> > > Would it be possible for the overlay to import code from `/front_end/` in that case?
> > 
> > If you are using Rollup anyways, I don't see why not.
> 
> I can try that, sound good. What would be specific maintenance benefits of having the inspector_overlay code on the upper level?

The fact that `front_end/` only contains concerning `front_end/` means we can write (for example) linter rules that codify invariants. Since the `/inspector_overlay/` does not (by default) adhere to these invariants, we would need to exempt it from many rules.

In any case, we are moving to a build system where code is grouped by purpose. To me, this CL once again points to the fact that `/inspector_overlay/` is a bit special (and it should be, since it is bundled separately in Chromium) and thus warrants its own space.

Patch-set: 8