Update patch set 11
Patch Set 11: Code-Review+1

(4 comments)

Nice, we are pretty much there! Let's make sure it keeps on working by adding it to the buildgraph and separate out the addition to `manage_node_deps.py`. The `node.py` change is more of an FYI and up to you if you want to invest additional time cleaning that up or leaving it as-is. I am okay with either choice.

Patch-set: 11
Reviewer: Gerrit User 1365394 <1365394@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+1