Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > Patch Set 3:
> > 
> > > Patch Set 3:
> > > 
> > > Hm, I don't think we can call `python` here, as it is local to your machine. If I understand correctly, Chromium is moving from Python 2 to 3, which requires the hashbang solution (I looked at https://bugs.chromium.org/p/chromium/issues/detail?id=942720 and linked CLs and they all use this solution). Can we find out what the Chromium guidance is for this transition with regards to Windows CMD? Can you run a different shell on Windows that would work?
> > 
> > Hmm, I think that the solution might be to call vpython instead? which I believe depot_tools is using to map to either the python 2 or 3 version bundled with depot_tools depending on where the migration is?
> 
> How do other Chromium contributors on Windows run Python scripts? I would prefer to use a standard solution. (If that is vpython, that's fine by me, but is there any documentation that states that vpython is the appropriate solution?)

I am not an expert by any means, and welcome others to comment. AFAIK, any of the following are acceptable: `python <script>` ; `vpython <script>` ; or `<script>` as part of checking out depot_tools/chromium, one of the steps is to ensure that any references to python point to depot_tools first, then the machine. I'm not sure what the "best practice" is though. 

I believe the particular error we are seeing is because the full path to node.py is not being recognized on windows, it thinks we are trying to execute a script called "third_party" and I don't think switching shells would resolve this issue. Brian can correct me if I'm wrong

Patch-set: 3