Update patch set 7
Patch Set 7:

> Patch Set 4:
> 
> > Patch Set 4:
> > 
> > > Patch Set 4:
> > > 
> > > I have honestly no idea if this is the right solution. Could you look into the two layout test failures? They seem related to the changes. Let's take it from there, after the tests are fixed, maybe you uncover another rabbit hole.
> > 
> > They are related and they will need to be disabled to check this in. One of them is simply due to expecting a snippet to be named with %escapes and the test will need to be fixed. The automapping test that Jan added I am less certain about. It appears to be testing against the local filesystem, but I am uncertain how to duplicate what it's doing either in manual testing or in an e2e test. Jan, can you shed any light on what that test is doing and how I might try to reproduce it manually?
> 
> The filesystem mapping is about this bug: crbug.com/992717
> If I remember correctly, automapping didn't work on windows when there were spaces in the filename. 
> This was due to things not being escaped properly.

Okay, I've figured out how to duplicate what the automapping test is testing and it appears I've regressed it. What is weird is that if you load a page using a file:/// url, the Page tab in sources will escape the url while the FileSystem tab for the same file will not escape the url. The page tab seems consistent between how it escapes file: urls and http: urls and is also consistent with getting those files by typing the urls into the chrome address bar, so I'd like to make the filesystem tab do escaping the same way, I'm just not sure where it is doing or should be doing that escaping.

Patch-set: 7
Label: Commit-Queue=0