Fix rollup inputs for restat computation
A Rollup action takes an entrypoint as input. However, we did not
specify this input explicitly in GN. This could then lead to
non-deterministic build failures, if a subsequent change to a
target using the `rollup_entrypoint` template would trigger a build.

To make sure that GN understands that the entrypoint is an input,
we have to add it to the `inputs` array of the target.

R=liviurau@chromium.org,aerotwist@chromium.org,jacktfranklin@chromium.org

Bug: 1011811
Change-Id: Idf03693575f68ea8fe06a2df10f895c4416bfe5c