Grid: Draw grid area names
This is based on this backend change:
https://chromium-review.googlesource.com/c/chromium/src/+/2275484

When the backend returns data about the named grid areas (their names
and paths), then the frontend of the overlay needs to draw them.
The mockup for how they should look can be found here:
https://docs.google.com/presentation/d/1krTiCT_c5sIAXBU6hpSok9ev3cZDRyucHf6rnKNTa8o/edit#slide=id.g74d98c0ef3_1_26

Grid areas are basically just highlighted with a thick border around
them, and the name is displayed in the top-left corner.
This change does not handle cases where multiple areas overlap,
resulting in overlapping names. This will be handled later.

Bug: 1100384
Change-Id: Icb935a1a6e143bfdd95babb3c66a8f1412c80418