Update patch set 1
Patch Set 1:

(1 comment)

Hi everyone,

This CL intends to make "npm run install-deps" work in Windows. This is still a WIP but want to be as cautious as possible to avoid breaking a specific platform. I tested this locally on Windows:

- No new dependencies added (just "npm run install-deps"): 
  works, but I get changes on several node_modules/.bin files that looks suspicious e.g:
  - going from ../mocha/bin/_mocha to a bash script (https://imgur.com/a/llYcYWv)
  - changes on "../mocha/bin/_mocha" but also a new "../mocha/bin/_mocha.cmd" being added.
- New dependencies added: 
  I'm getting an error related to package.json and package.json.lock out of sync. Still investigating if this is related to my environment settings.


As a proof of concept, I ran the script with some workarounds and got this:
(added rollup-commonjs)
https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2277235

Please let me know your thoughts.

Patch-set: 1
CC: Gerrit User 1341658 <1341658@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
CC: Gerrit User 1341669 <1341669@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
CC: Gerrit User 1365394 <1365394@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Work-in-progress: false