Fix Network tab request sorting for "path" and "url" columns
Currently the "sort by column" functionality does not work for "path"
and "url" columns in the Network tab. This is because the comparator
expects a property name in a NetworkRequest, while "path" and "url"
are method names. There are several options to fix this:
1. change "path" and "url" to getters, but this will break existing
behaviors for sure;
2. in the current RequestPropertyComparator used for "path" and "url",
we add checks to see if the property is in fact a method, thus invoking
the method. This is not ideal because it "assumes" too much custom logic
as a low-level utility function, and opening this door means future
use cases of mixing method names and property names are also valid.
I don't think this mixing is good generally;
3. add a separate function just for url comparison. This has the
downside of adding more code, but I think the benefit of expressiveness
and simplicity justifies the cost. Also there are already several
comparators that are created for this reason (e.g. remoteAddress). I
chose this option for now.

Separately, "path()" is not even the correct method the sort should base
on, because the "Path" column actually shows "pathname" property values,
which is different from "path()".

Bug: chromium:1090191
Change-Id: I85a1de6385c0a96842dc68af8ffe9531bbf55ed9