Add experimental new frame tree to application panel
This CL adds a tree for visualising a page's iframe structure in the
application panel. In this CL the focus lies on correctly displaying
the frames (naming, updating upon page navigation) and highlighting
the respective frames in the main browser window when hovering over
their names in the frame tree.

This new frame tree will replace or be merged with the existing
frame tree, which is buggy, but shows frame resources as well.

Screenshot: https://i.imgur.com/GMqPQiE.png

Bug: chromium:1093247, chromium:1101262

Change-Id: Iaa0a01ee4e367b2fcee00772daf4c95069c3e65b