Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> The changes are a bit weird to me, especially the changes to the various package.json files. What version of NPM are you running?
npm --version = 6.14.5

I agree the changes in the binary folder also look strange, and if I compare this to the one you created:
2279833: Add @rollup/plugin-commonjs to NPM dependencies | https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2279833

The number of changed files is also really odd. My eyes are set on:

rm -rf node_modules && npm install

is not the same as (even under specific circunstances) to:

npm ci 

So, I'll continue investigation.

Patch-set: 2
Label: Commit-Queue=0
Work-in-progress: false