Fix disappearing local filesystems when group mode is changed
When the setting "Group by folder" is changed, the entire sources UI
is rerendered. Rerendering removes filesystem subfolder nodes that are
added based on the events from the underlying workspace. Therefore,
the local filesystem didn't render after you change the setting. This
CL adds a call to _resetWorkspace to re-fetch workspace folders when
the UI is reset.

Fixed: chromium:829319
Bug: chromium:1066415
Change-Id: I0cbb481f26a7f9bb747b0ed608b6cee2bdbd856c