Remove unused getter on Cookie
The port attribute was only allowed for Set-Cookie2, but we never
surfaced that information in our structured cookie display logic.

From a UX perspective, the port is part of the domain, and should be
communicated as such, so it doesn't make a lot of sense to keep this
around.

Fixed: chromium:1060613
Change-Id: Ia4c23d47604185f55ef99377d5af2e7d073d97cd