Fix generated tsconfig in devtools_entrypoint
Before this change, we would incorrectly rename the generated tsconfig
as the second step of the devtools_entrypoint template. This would
then make TypeScript upset, as it was not able to find the built
Rollup bundle in release mode.

To fix that issue, we have to generate a custom `tsconfig.json`
(very much alike what we do for the third_party packages) and make
sure that the entrypoint has a corresponding `.d.ts` file as well.

CC=jacktfranklin@chromium.org

Bug: 1098730, 1101738, 1061037
Change-Id: I25e79e6c9d28377a752212c7380627d1c32ebb97
Also-By: aerotwist@chromium.org