test: Use root hook plugins for Mocha
Mocha 8 introduces 'root hook plugins' which is an alternative way of
specifying root hooks, which is required for parallel mode.

This CL uses the new root hook plugins following the migration guide:
https://mochajs.org/#migrating-tests-to-use-root-hook-plugins

This CL does not make us ready for parallel mode yet, because we still
have a few assumptions that we have one instance running. These will be
dealt with in future CLs. This CL just ports the root hooks over to the
plugin system while maintaining the same behavior.

Bug: 1101784
Change-Id: I74dfc7f24decd031ebac656c7cea8ef87f9ac791