Disable flaky e2e tests that are blocking trybots
https://logs.chromium.org/logs/devtools-frontend/buildbucket/cr-buildbucket.appspot.com/8875541506718258480/+/steps/E2E_tests/0/stdout
https://logs.chromium.org/logs/devtools-frontend/buildbucket/cr-buildbucket.appspot.com/8875542083664424752/+/steps/E2E_tests/0/stdout
https://logs.chromium.org/logs/devtools-frontend/buildbucket/cr-buildbucket.appspot.com/8875543132396125056/+/steps/E2E_tests/0/stdout
https://logs.chromium.org/logs/devtools-frontend/buildbucket/cr-buildbucket.appspot.com/8875543394161054208/+/steps/E2E_tests/0/stdout

These tests have failed at least 4 times this morning already, and with
the 2+ hour queue time on the mac bot, mean that code can't be landed.

This CL has to skip presubmit because there is a broke presubmit
that a skipped issue has a crbug in the name, which doesn't work for
template strings.

No-Presubmit: true
No-Tree-Checks: true
No-Try: true
Bug: 1073406
Change-Id: Ib54f60d49968f323aa4dc76020443442f54e52d9