tests: Pass the ports we are using to the hosted mode server
Right now we hardcode the ports and assume that the hosted mode server
hardcodes the same ones. In the future we will have different port
numbers for different parallel test runners, so we need to pass the
correct numbers here.

env.PORT is the hosted mode server port, so it's essential we get that
right. env.REMOTE_DEBUGGING_PORT is only used in a link on the
landing page of the hosted mode server, so it being correct is not
essential for the tests to run properly but there's no reason for
it to be wrong.

Also rename envPort to chromeDebugPort to make things clearer with all
the different port numbers flying around.

Bug: 1101784
Change-Id: I58880a3543021bbcd484cf3578ff8225d3e3d81c