Make `devtools_module` visible only to targets in the same folder
To make sure that someone doesn't accidentally depend on the internal
implementation of a module (e.g. a `devtools_module` target), we can
set the target visibility to `:*`. This is equivalent to "all targets
in the same file".

This resulted in some build failures, as indeed there were a couple
of wrong build dependencies, which should have depended on `:bundle`
instead.

Consequently, because the targets now are no longer publicly visible,
GN complains that the GRD task at the end can not find the files
they generate. To remedy that problem, we have to add most of the
actions to `public_deps`. This is not a problem per se, as these
actions are internal implementations anyways and not visible.

DISABLE_THIRD_PARTY_CHECK=Build fixes for TypeScript
R=aerotwist@chromium.org

Bug: 1101738
Change-Id: I36e4081797fe0f67be584f0ee15c1533a3449ea2