tests: Wait for the hosted mode server to start
Previously we started the hosted mode server and then
started chrome without waiting for it to start. This is
probably fine in most situations because chrome takes a
while to start and we need to do extra setup there, so
we have a long time before the server needs to respond.

In the server itself, we add two messages that the server
will send over IPC if the channel is open: PORT_IN_USE
and READY. Mocha can now wait for either of these two
events before going ahead and starting chrome.

The reason we want to do this is because with the mocha
parallel mode, we may try to start the server with a
port that is already in use, and we need a way to
distinguish these errors from other failures of the
hosted mode server. We also want to retry starting the
server with a different port in that situation, and
eventually give up if something is wrong. This also
makes the mocha runner more robust to failures e.g.
if we accidentally leave the server running and the
port is taken.

This doesn't slow down the startup time of the e2e
test runner at all, according to timing of local runs.

Bug: 2284839
Change-Id: Ibb14d5a177ba5f9e257743824770bff3d5ccb033