Create console UI async
To fix 696805 we need to build console UI in an async way to allow
querying remote objects when we format template strings. This CL
makes UI creation functions async and extracts the logic from the
getters. A follow-up will contain the actual
fix involving invoking toString method of the remote object when
the first argument to console message is a template string.

Bug: 696805

Change-Id: I4f17263791ef7fd3cda1783174b8978c30204058