Add support for `skip_rollup` in `module.json`
For folders that already have a `devtools_entrypoint` target, we
should skip Rollup as part of `build_release_applications.py`. To
remove the hardcoded list of skipped folders and enable
parallelization of a fixit to migrate all folders, we have to put
the information in the `module.json` instead.

Once all folders are migrated to `devtools_entrypoint`, we can remove
the field and the Rollup logic in `build_release_applications.py`.

R=aerotwist@chromium.org

Bug: 1101738
Change-Id: Ibadab13e87cf28622c40896a56f6ed18cc34bea2